% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenem.R
\name{catenem}
\alias{catenem}
\title{Faz a limpeza do banco de dados do Enem}
\usage{
catenem(dados)
}
\arguments{
\item{dados}{banco de dados gerado apos a limpeza}
}
\value{
o banco de dados do ano \code{ano}.
}
\description{
Funcao para fazer a limpeza e organizacao dos
microdados do Enem
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
    sua funcao (se necessario), ou para detalhar melhor como
    utilizar determinados argumentos.
}
\examples{
soma(2, 2)

x <- 3
y <- 4
soma(x = x, y = y)

}
\author{
Fernando Bastos
}
