% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{dados}
\alias{dados}
\title{dados.}
\format{
A data frame with 27 rows and 3 variables:
\describe{
  \item{ESTADO}{Acronym of Brazilian States}
  \item{POPULACAO}{Population of Brazilian States.}
  \item{REGIAO}{Region of Brazilian States.}
  ...
}
}
\usage{
dados
}
\description{
data frame 2018.
}
\keyword{datasets}
