% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enemclear.R
\name{enemclear}
\alias{enemclear}
\title{Makes the cleaning of the Enem database}
\usage{
enemclear(year)
}
\arguments{
\item{year}{Year you want the microdata}
}
\value{
the database of the year \code{year}.
}
\description{
Function to do the cleaning and organization of the
Enem microdata
}
\author{
Fernando Bastos
}
