% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\docType{data}
\name{pop}
\alias{pop}
\title{Pop.}
\format{
A data frame with 27 rows and 3 variables:
\describe{
  \item{ESTADO}{Acronym of Brazilian States}
  \item{POPULACAO}{Population of Brazilian States.}
  \item{REGIAO}{Region of Brazilian States.}
  ...
}
}
\usage{
pop
}
\description{
Population of Brazilian States.
}
\keyword{datasets}
