% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readenem.R
\name{readenem}
\alias{readenem}
\title{Faz a leitura do banco de dados do Enem}
\usage{
readenem(ano)
}
\arguments{
\item{ano}{Ano que voce deseja os microdados}
}
\value{
o banco de dados do ano \code{ano}.
}
\description{
Funcao para fazer a leitura dos
microdados do Enem
}
\details{
Utilize este campo para escrever detalhes mais tecnicos da
    sua funcao (se necessario), ou para detalhar melhor como
    utilizar determinados argumentos.
}
\examples{
soma(2, 2)

x <- 3
y <- 4
soma(x = x, y = y)

}
\author{
Fernando Bastos
}
