% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Compiles the shiny}
\usage{
runShinyApp()
}
\value{
An interactive dashboard.
}
\description{
Function to generate the shiny
}
\author{
Fernando de Souza Bastos
}
