% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enemdown.R
\name{enemdown}
\alias{enemdown}
\title{Download the Enem database}
\usage{
enemdown(year)
}
\arguments{
\item{year}{Year you want the microdata}
}
\value{
the database of the year \code{year}.
}
\description{
Function to download the
Enem microdata
}
\author{
Fernando de Souza Bastos
}
